removed binaries

+++ dist.sh
@@ -1,7 +1,13 @@
 #!/bin/bash
 
-rm -r dist/*
-rm HaxeUmlGen.zip
+if [ -e dist ]; then
+    rm -r dist/*
+fi
+
+if [ -e HaxeUmlGen.zip ]; then
+    rm HaxeUmlGen.zip
+fi
+
 cp -r haxelib.xml bin/* LICENSE README dist
 cd src
 cp -r umlgen ../dist
